[REFACTOR] 기존의 MealPlanResult 컴포넌트를 meal-plan-create-page파일에 추가
